fix: revue emprise administrative, migration de backfill et corrections
Suites de la revue de la branche.

Déploiement : une recherche « Bretagne » aurait renvoyé zéro résultat entre la
MEP et le nightly de 6 h (champs non peuplés, index Atlas Search non reconstruit).
La migration 20260911180000 applique la définition de l'index (updateSearchIndex,
idempotent) et met fillSearchItemsCollection en file pour le job processor.

Nightly : un seul chemin par source, item nouveau ou déjà indexé, via
upsertSearchItem qui réécrit tout sauf keywords. L'ancien $set d'une liste fermée
de champs pour les items existants est ce qui a laissé dériver address et
location (1 593 offres en prod). Une formation sans géopoint est comptée
(formations_sans_geopoint) plutôt que signalée par exception.

Distance : masquée quand admin_area est posé, elle aurait été mesurée depuis le
centroïde de la région. Le tri par proximité garde lat/lon.

Télémétrie : admin_area dans search_queries et search_admin_area dans les
événements Matomo search_launched et search_results_displayed.

Regex admin_area et parseur partagés dans shared/utils/admin-area (route, moteur,
URL de la page). Deux exports sans usage externe retirés (knip).
