+++ server/src/services/geo-administrative/admin-area.ts
@@ -1,3 +1,5 @@
+import type { IAdminArea } from "shared/utils/admin-area"
+
 /**
  * Filtrage de la recherche par entité administrative (région, département).
  * Contexte et mesures : issue BaseAdresseNationale/ban-plateforme#781, tools/geo-781.
@@ -8,23 +10,8 @@
  * `token` : le filtre est un `equals`, exact et le moins coûteux possible.
  */
 
-export type IAdminArea = {
-  kind: "region" | "departement"
-  code: string
-}
-
-const ADMIN_AREA_PATTERN = /^(region|departement):([0-9]{1,3}|2[AB])$/
-
-/** `"region:53"` / `"departement:44"`. Renvoie null sur toute autre forme, jamais d'exception. */
-export function parseAdminArea(value: string | undefined | null): IAdminArea | null {
-  if (!value) return null
-  const match = ADMIN_AREA_PATTERN.exec(value)
-  if (!match) return null
-  return { kind: match[1] as IAdminArea["kind"], code: match[2] }
-}
-
 /** Champ `search_items` porteur du code pour la maille demandée. */
-export const ADMIN_AREA_FIELD: Record<IAdminArea["kind"], "departement_code" | "region_code"> = {
+const ADMIN_AREA_FIELD: Record<IAdminArea["kind"], "departement_code" | "region_code"> = {
   region: "region_code",
   departement: "departement_code",
 }
