+++ server/src/services/search/search-query-log.service.ts
@@ -47,6 +47,7 @@ type SearchQuerystring = {
   latitude?: number
   longitude?: number
   radius?: number
+  admin_area?: string
   search_source?: "suggestion" | "free_text" | "training_links" | "external_sites"
   // Alias déprécié (cf. search.routes.ts) : encore envoyé volontairement par l'UI courante
   // (transition anti-version-skew) et par les liens traininglinks émis avant le 2026-08-24.
@@ -89,6 +90,7 @@ export async function logSearchQuery(query: SearchQuerystring, outcome: SearchOu
       has_geo: hasGeo,
       geo: hasGeo ? { lat: roundCoord(query.latitude!), lng: roundCoord(query.longitude!) } : null,
       radius: hasGeo ? (query.radius ?? null) : null,
+      admin_area: query.admin_area ?? null,
       created_at: now,
     })
   } catch (error) {
